+++ .github/workflows/deploy_website_content.yml
@@ -78,6 +78,7 @@ jobs:
           export NEXT_PUBLIC_redirectSignIn="https://${fullCloudfrontDomain}/site/auth_demo.html"
           export NEXT_PUBLIC_redirectSignOut="https://${fullCloudfrontDomain}/site/"
           export NEXT_PUBLIC_COMMIT_ID=${{ inputs.COMMIT_ID }}
+          export NEXT_PUBLIC_ENV=${{ inputs.TARGET_ENVIRONMENT }}
 
           cd .build
           make react-build
