+++ packages/cdk/resources/Cognito.ts
@@ -188,20 +188,20 @@ export class Cognito extends Construct {
     const callbackUrls = [
       `https://${props.fullCloudfrontDomain}/site/`,
       // FIXME: This is temporary, until we get routing fixed
-      `https://${props.fullCloudfrontDomain}/site/auth_demo.html`,
-      `https://${props.fullCloudfrontDomain}/auth_demo/`,
+      `https://${props.fullCloudfrontDomain}/site/login.html`,
+      `https://${props.fullCloudfrontDomain}/login/`,
       `https://${props.fullCloudfrontDomain}/oauth2/idpresponse`
     ]
 
     const logoutUrls = [
       `https://${props.fullCloudfrontDomain}/site/`,
-      `https://${props.fullCloudfrontDomain}/site/auth_demo.html`,
-      `https://${props.fullCloudfrontDomain}/auth_demo/`
+      `https://${props.fullCloudfrontDomain}/site/login.html`,
+      `https://${props.fullCloudfrontDomain}/login/`
     ]
 
     if (props.useLocalhostCallback) {
       callbackUrls.push( "http://localhost:3000/auth/")
-      callbackUrls.push( "http://localhost:3000/auth_demo/")
+      callbackUrls.push( "http://localhost:3000/login/")
       logoutUrls.push( "http://localhost:3000/")
     }
     // add the web client
